# Local mtime check